refactor: EducationSection component📦
